requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
requests==2.32.5
feedparser==6.0.12
newspaper3k==0.2.8
beautifulsoup4==4.13.5
lxml==6.0.1
nltk==3.9.1
schedule==1.2.2
pydantic==2.11.7
python-multipart==0.0.12